[update] Changed the name of the project to productivity tools and converted days since into a project
