+++ days_since/apps.py
@@ -1,5 +1,5 @@
 from django.apps import AppConfig
 
 
-class WebConfig(AppConfig):
+class DaysSinceConfig(AppConfig):
     name = 'days_since'
