+++ days_since/templates/days_since/activity_detail.html
@@ -5,7 +5,7 @@
 {% block content %}
 <div class="ui container">
 	<div class="ui hidden divider"></div>
-	<a href="{% url 'web:home' %}" class="ui button">
+	<a href="{% url 'days_since:home' %}" class="ui button">
 		<i class="ui chevron left icon"></i>
 		Back to Home
 	</a>
@@ -13,17 +13,17 @@
 	<h1>{{activity.title}}</h1>
 	<p>{{activity.description}}</p>
 	<p>Frequency: {{activity.frequency_display}}</p>
-	<a href="{% url 'web:update-activity' pk=activity.pk %}" class="ui teal button">
+	<a href="{% url 'days_since:update-activity' pk=activity.pk %}" class="ui teal button">
 		<i class="ui edit icon"></i>
 		Edit
 	</a>
-	<a href="{% url 'web:delete-activity' pk=activity.pk %}" class="ui red button">
+	<a href="{% url 'days_since:delete-activity' pk=activity.pk %}" class="ui red button">
 		<i class="ui trash icon"></i>
 		Delete
 	</a>
 	<div class="ui hidden divider"></div>
 	<h2>Events</h2>
-	<a href="{% url 'web:create-event' activity=activity.pk %}" class="ui blue button">
+	<a href="{% url 'days_since:create-event' activity=activity.pk %}" class="ui blue button">
 		<i class="ui plus icon"></i>
 		Create Event
 	</a>
@@ -32,10 +32,10 @@
 			<div style="display: flex; align-items: center; justify-content: space-between;">
 				<div>{{event.get_event_type_display}} on {{event.date}}</div>
 				<div>
-					<a href="{% url 'web:update-event' pk=event.pk %}" class="ui small teal icon button">
+					<a href="{% url 'days_since:update-event' pk=event.pk %}" class="ui small teal icon button">
 						<i class="ui edit icon"></i>
 					</a>
-					<a href="{% url 'web:delete-event' pk=event.pk %}" class="ui small red icon button">
+					<a href="{% url 'days_since:delete-event' pk=event.pk %}" class="ui small red icon button">
 						<i class="ui trash icon"></i>
 					</a>
 				</div>
