+++ days_since/templates/days_since/delete_form.html
