+++ productivity_tools/asgi.py
@@ -1,5 +1,5 @@
 """
-ASGI config for days_since project.
+ASGI config for productivity_tools project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'days_since.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'productivity_tools.settings')
 
 application = get_asgi_application()
