+++ productivity_tools/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for days_since project.
+Django settings for productivity_tools project.
 
 Generated by 'django-admin startproject' using Django 3.0.3.
 
@@ -24,7 +24,7 @@ BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 SECRET_KEY = 'kbx4jw9+3xhipkw@3g@slh*3&z*a^8jr+-!n%v+z(ul$+dy4_g'
 
 # SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = False
+DEBUG = True
 
 ALLOWED_HOSTS = []
 
@@ -44,9 +44,10 @@ INSTALLED_APPS = [
     'semanticuiforms',
     'timezone_field',
 
+    'registration.apps.RegistrationConfig',
+    'days_since',
     'api',
     'web',
-    'registration.apps.RegistrationConfig',
 ]
 
 MIDDLEWARE = [
@@ -60,7 +61,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'days_since.urls'
+ROOT_URLCONF = 'productivity_tools.urls'
 
 TEMPLATES = [
     {
@@ -78,7 +79,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'days_since.wsgi.application'
+WSGI_APPLICATION = 'productivity_tools.wsgi.application'
 
 
 # Database
