+++ templates/rest_framework/api.html
@@ -3,6 +3,6 @@
 
 {% block branding %}
     <a class='navbar-brand' rel="nofollow" href='/api' %}'>
-        Days Since API
+        Productivity Tools API
     </a>
 {% endblock %}
